+++ integration-cli/docker_cli_build_test.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/docker/docker/archive"
+	"github.com/docker/docker/pkg/archive"
 )
 
 func TestBuildCacheADD(t *testing.T) {
