+++ pkg/archive/diff.go
